+++ QBCore_2F0666.base/resources/[qb]/qb-multicharacter/server.lua
@@ -163,7 +163,7 @@ local function GiveStarterItems(source)
             info.firstname = Player.PlayerData.charinfo.firstname
             info.lastname = Player.PlayerData.charinfo.lastname
             info.birthdate = Player.PlayerData.charinfo.birthdate
-            info.type = 'Class C Driver License'
+            info.type = 'B Sınıfı Sürücü Belgesi'
         end
         exports['qb-inventory']:AddItem(src, v.item, v.amount, false, info, 'qb-multicharacter:GiveStarterItems')
     end