+++ QBCore_2F0666.base/resources/[qb]/qb-smallresources/client/cruise.lua
@@ -30,17 +30,10 @@ local function triggerCruiseControl(veh)
         if speed > 0 and GetVehicleCurrentGear(veh) > 0 then
             speed = GetEntitySpeed(veh)
             local isTurningOrHandbraking = IsControlPressed(2, 76) or IsControlPressed(2, 63) or IsControlPressed(2, 64)
-            if GetResourceState('qb-fuel') == 'started' then
-                if exports['qb-fuel']:GetFuel(veh) <= 10 then
+            if GetResourceState('qb-fuel') == 'started' and exports['qb-fuel']:GetFuel(veh) <= 10 then
                     QBCore.Functions.Notify(Lang:t('cruise.not_Enough_Fuel'), 'error')
                     return
                 end
-            elseif GetResourceState('LegacyFuel') == 'started' then
-                if exports['LegacyFuel']:GetFuel(veh) <= 10 then
-                    QBCore.Functions.Notify(Lang:t('cruise.not_Enough_Fuel'), 'error')
-                    return
-                end
-            end
             TriggerEvent('seatbelt:client:ToggleCruise', true)
             QBCore.Functions.Notify(Lang:t('cruise.activated'))
             CreateThread(function()
@@ -85,4 +78,4 @@ RegisterCommand('togglecruise', function()
     end
 end, false)
 
-RegisterKeyMapping('togglecruise', 'Toggle Cruise Control', 'keyboard', 'Y')
+RegisterKeyMapping('togglecruise', 'Hız Sabitleyici', 'keyboard', 'Y')