+++ QBCore_2F0666.base/resources/[qb]/qb-vehiclekeys/fxmanifest.lua
@@ -13,7 +13,10 @@ shared_scripts {
 }
 
 client_script 'client.lua'
-server_script 'server.lua'
+server_scripts {
+    '@oxmysql/lib/MySQL.lua',
+    'server.lua',
+}
 
 ui_page 'NUI/index.html'