+++ QBCore_2F0666.base/resources/[standalone]/rp-dealership/config.lua
@@ -34,7 +34,7 @@ Config.CategoryLabels = {
     military = 'Askeri',
     commercial = 'Ticari',
     trains = 'Tren',
-    openwheel = 'Open Wheel',
+    openwheel = 'Açık Teker',
 }
 
 Config.Dealerships = {